Extract shared Vectorize primitives out of the MCP layer
The package registry, community search, and job reindexing all embedded
and queried the shared CAPABILITY_VECTOR_INDEX binding by importing
#mcp/capabilities/capability-search.ts, which made non-MCP subsystems
depend on the MCP layer for generic infrastructure.

Workers AI embedding, index access, offline detection, hybrid ranking
primitives, length-safe vector ids, and reindex batching now live under
#worker/vectorize/*. capability-search.ts keeps only capability-specific
search (builtin kind filter, embed text, capability ranking rows).

Embedding-wrapper tests move alongside the module they cover.
