+++ packages/worker/src/mcp/memory/memory-reindex.node.test.ts
@@ -13,7 +13,7 @@ const mockModule = vi.hoisted(() => ({
 	listMemoriesPage: vi.fn(),
 }))
 
-vi.mock('#mcp/capabilities/capability-search.ts', () => ({
+vi.mock('#worker/vectorize/embedding.ts', () => ({
 	embedTextForVectorize: (...args: Array<unknown>) =>
 		mockModule.embedTextForVectorize(...args),
 	embedTextsForVectorize: (...args: Array<unknown>) =>
