+++ packages/worker/src/package-registry/package-reindex-d1-retry.node.test.ts
@@ -16,7 +16,7 @@ const mockModule = vi.hoisted(() => ({
 	loadPublishedEntityManifest: vi.fn(),
 }))
 
-vi.mock('#mcp/capabilities/capability-search.ts', () => ({
+vi.mock('#worker/vectorize/embedding.ts', () => ({
 	embedTextsForVectorize: (...args: Array<unknown>) =>
 		mockModule.embedTextsForVectorize(...args),
 	getCapabilityVectorIndex: (...args: Array<unknown>) =>
