+++ packages/worker/src/vectorize/vector-ids.node.test.ts
